+++ packages/integrity/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @meshql/core@0.5.1
+  - @meshql/http@0.5.1
+
 ## 0.5.0
 
 ### Minor Changes